requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
psycopg2-binary==2.9.9
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.0
pypdf==4.3.1
requests==2.32.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
